+++ app/src/main/java/com/github/welcomeworld/bangumi/instrumentality/project/source/bili/retrofit/databean/IndexRecommendBean.java
@@ -4,16 +4,16 @@ import java.util.List;
 
 @SuppressWarnings("unused")
 public class IndexRecommendBean {
-    private int code;
+    private long code;
     private ConfigBean config;
     private String message;
     private List<IndexRecommendDataBean> data;
 
-    public int getCode() {
+    public long getCode() {
         return code;
     }
 
-    public void setCode(int code) {
+    public void setCode(long code) {
         this.code = code;
     }
 
@@ -43,13 +43,13 @@ public class IndexRecommendBean {
 
     public static class ConfigBean {
 
-        private int feed_clean_abtest;
+        private long feed_clean_abtest;
 
-        public int getFeed_clean_abtest() {
+        public long getFeed_clean_abtest() {
             return feed_clean_abtest;
         }
 
-        public void setFeed_clean_abtest(int feed_clean_abtest) {
+        public void setFeed_clean_abtest(long feed_clean_abtest) {
             this.feed_clean_abtest = feed_clean_abtest;
         }
     }
